Helpers: group resource fetching and font cache ownership
