+++ takumi-helpers/src/fonts.ts
@@ -1,15 +1,7 @@
 import type { GoogleFontShapeFamilies, GoogleFontShapes } from "./google-fonts-catalog";
 import type { Node } from "./types";
-import {
-  defaultMaxFetchBytes,
-  fetchOk,
-  type FetchLike,
-  type FetchOptions,
-  readBodyLimited,
-} from "./utils";
-
-const chromeUserAgent =
-  "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36";
+import { defaultMaxFetchBytes, fetchOk, type FetchOptions, readBodyLimited } from "./fetch";
+import { FontStylesheetCache } from "./fonts/cache";
 
 type FontStyle = "normal" | "italic";
 
@@ -85,76 +77,6 @@ export type GoogleFontsOptions = FetchOptions & {
 
 const GOOGLE_FONTS_CSS = "https://fonts.googleapis.com/css2";
 
-const maxCachedCssBytes = 1024 * 1024;
-const maxCachedCssEntries = 64;
-
-class FontCssCache {
-  private static readonly byFetch = new WeakMap<FetchLike, FontCssCache>();
-  private readonly entries = new Map<string, { pending: Promise<string>; bytes: number }>();
-  private bytes = 0;
-
-  static forFetch(fetch: FetchLike) {
-    let cache = this.byFetch.get(fetch);
-    if (!cache) {
-      cache = new FontCssCache();
-      this.byFetch.set(fetch, cache);
-    }
-    return cache;
-  }
-
-  get(url: string) {
-    const entry = this.entries.get(url);
-    if (entry) {
-      this.entries.delete(url);
-      this.entries.set(url, entry);
-    }
-    return entry?.pending;
-  }
-
-  set(url: string, pending: Promise<string>) {
-    this.delete(url);
-    const entry = { pending, bytes: url.length * 2 };
-    this.entries.set(url, entry);
-    this.bytes += entry.bytes;
-    this.trim();
-    pending.then(
-      (css) => {
-        if (this.entries.get(url) !== entry) {
-          return;
-        }
-        entry.bytes += css.length * 2;
-        this.bytes += css.length * 2;
-        if (entry.bytes > maxCachedCssBytes) {
-          this.delete(url);
-        }
-        this.trim();
-      },
-      () => {
-        if (this.entries.get(url) === entry) {
-          this.delete(url);
-        }
-      },
-    );
-  }
-
-  delete(url: string) {
-    const entry = this.entries.get(url);
-    if (entry) {
-      this.bytes -= entry.bytes;
-      this.entries.delete(url);
-    }
-  }
-
-  private trim() {
-    for (const url of this.entries.keys()) {
-      if (this.entries.size <= maxCachedCssEntries && this.bytes <= maxCachedCssBytes) {
-        break;
-      }
-      this.delete(url);
-    }
-  }
-}
-
 // Builds a css2 `family=` value, e.g. `Inter:ital,opsz,wght@0,14..32,400`. Takes plain values
 // rather than a GoogleFontFamily so it never relates that ~2000-member union to a parameter, which
 // overflows the type checker (TS2590).
@@ -212,48 +134,6 @@ function buildUrl(options: GoogleFontsOptions) {
     url.searchParams.set("display", options.display);
   }
   return url.toString();
-}
-
-const maxCssBytes = 2 * 1024 * 1024;
-
-function fetchCss(url: string, options: FetchOptions) {
-  return fetchOk(url, {
-    ...options,
-    init: { headers: { "User-Agent": chromeUserAgent } },
-  })
-    .then((r) => readBodyLimited(r, options.maxBytes ?? maxCssBytes))
-    .then((buffer) => new TextDecoder().decode(buffer));
-}
-
-function fetchCssCached(url: string, options: GoogleFontsOptions) {
-  options.signal?.throwIfAborted();
-  if (options.signal || options.allowUrl) {
-    return fetchCss(url, options);
-  }
-  const cache = options.cache ?? FontCssCache.forFetch(options.fetch ?? globalThis.fetch);
-
-  const cached = cache.get(url);
-  if (cached) {
-    return cached.then((css) => {
-      const maxBytes = options.maxBytes ?? maxCssBytes;
-      if (new TextEncoder().encode(css).byteLength > maxBytes) {
-        throw new Error(`Response exceeds ${maxBytes} bytes`);
-      }
-      return css;
-    });
-  }
-
-  const pending = fetchCss(url, options);
-  cache.set(url, pending);
-  if (options.cache) {
-    pending.catch(() => {
-      if (cache.get(url) === pending) {
-        cache.delete(url);
-      }
-    });
-  }
-
-  return pending;
 }
 
 const fontFaceBlockPattern = /(?:\/\*\s*([^*]+?)\s*\*\/\s*)?@font-face\s*\{([^}]*)\}/g;
@@ -532,7 +412,7 @@ export async function googleFonts(
     return [];
   }
 
-  const css = await fetchCssCached(buildUrl(options), options);
+  const css = await FontStylesheetCache.load(buildUrl(options), options);
 
   // Google's css2 response orders `@font-face` blocks by its own logic, not by the `family=`
   // query order — the caller's declared priority survives only if we restore it here. A render
