+++ takumi-helpers/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./fonts";
 export * from "./helpers";
 export * from "./types";
-export * from "./utils";
+export * from "./fetch";
+export * from "./images";
